+++ src/plugins/intel_npu/src/backend/include/zero_profiling.hpp
