+++ src/plugins/intel_npu/src/compiler/include/iexternal_compiler.hpp
